# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PM Network API - GitHub for Product Managers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  followers?: number;
  following?: number;
  deckCount?: number;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 2 */
  username: string;
  /** @minLength 1 */
  name: string;
  bio?: string;
  title?: string;
  company?: string;
  avatarUrl?: string;
}

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  bio?: string;
  title?: string;
  company?: string;
  avatarUrl?: string;
}

export interface FollowInput {
  followerId: number;
}

export interface FollowResult {
  following: boolean;
  followerCount?: number;
}

export interface Deck {
  id: number;
  userId: number;
  authorName?: string;
  authorUsername?: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  tags?: string[];
  stars: number;
  views: number;
  createdAt: string;
  updatedAt?: string;
}

export interface DeckInput {
  userId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  thumbnailUrl?: string;
  fileUrl?: string;
  tags?: string[];
}

export interface DeckUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  thumbnailUrl?: string;
  fileUrl?: string;
  tags?: string[];
}

export interface StarInput {
  userId: number;
}

export interface StarResult {
  starred: boolean;
  starCount: number;
}

export type CourseLevel = typeof CourseLevel[keyof typeof CourseLevel];


export const CourseLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  instructor?: string | null;
  provider: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  duration?: string | null;
  level: CourseLevel;
  tags?: string[];
  isFree?: boolean;
  createdAt: string;
}

export type CourseInputLevel = typeof CourseInputLevel[keyof typeof CourseInputLevel];


export const CourseInputLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface CourseInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  instructor?: string;
  /** @minLength 1 */
  provider: string;
  url?: string;
  thumbnailUrl?: string;
  duration?: string;
  level: CourseInputLevel;
  tags?: string[];
  isFree?: boolean;
}

export type EventType = typeof EventType[keyof typeof EventType];


export const EventType = {
  case_comp: 'case_comp',
  pm_event: 'pm_event',
  workshop: 'workshop',
  conference: 'conference',
} as const;

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  type: EventType;
  /** @nullable */
  host?: string | null;
  eventDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  registrationDeadline?: string | null;
  /** @nullable */
  prize?: string | null;
  tags?: string[];
  createdAt: string;
}

export type EventInputType = typeof EventInputType[keyof typeof EventInputType];


export const EventInputType = {
  case_comp: 'case_comp',
  pm_event: 'pm_event',
  workshop: 'workshop',
  conference: 'conference',
} as const;

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  type: EventInputType;
  host?: string;
  eventDate: string;
  endDate?: string;
  location?: string;
  url?: string;
  registrationDeadline?: string;
  prize?: string;
  tags?: string[];
}

export type EventUpdateType = typeof EventUpdateType[keyof typeof EventUpdateType];


export const EventUpdateType = {
  case_comp: 'case_comp',
  pm_event: 'pm_event',
  workshop: 'workshop',
  conference: 'conference',
} as const;

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  type?: EventUpdateType;
  host?: string;
  eventDate?: string;
  endDate?: string;
  location?: string;
  url?: string;
  registrationDeadline?: string;
  prize?: string;
  tags?: string[];
}

export interface PlatformStats {
  userCount: number;
  deckCount: number;
  starCount: number;
  courseCount: number;
  eventCount: number;
  upcomingEventCount?: number;
}

export type ListUsersParams = {
search?: string;
limit?: number;
offset?: number;
};

export type ListDecksParams = {
search?: string;
tag?: string;
sortBy?: ListDecksSortBy;
limit?: number;
offset?: number;
};

export type ListDecksSortBy = typeof ListDecksSortBy[keyof typeof ListDecksSortBy];


export const ListDecksSortBy = {
  stars: 'stars',
  views: 'views',
  newest: 'newest',
} as const;

export type GetTrendingDecksParams = {
limit?: number;
};

export type ListCoursesParams = {
search?: string;
level?: ListCoursesLevel;
tag?: string;
limit?: number;
offset?: number;
};

export type ListCoursesLevel = typeof ListCoursesLevel[keyof typeof ListCoursesLevel];


export const ListCoursesLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type ListEventsParams = {
search?: string;
type?: ListEventsType;
upcoming?: boolean;
limit?: number;
offset?: number;
};

export type ListEventsType = typeof ListEventsType[keyof typeof ListEventsType];


export const ListEventsType = {
  case_comp: 'case_comp',
  pm_event: 'pm_event',
  workshop: 'workshop',
  conference: 'conference',
} as const;

export type GetUpcomingEventsParams = {
limit?: number;
};