# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * PM Network API - GitHub for Product Managers
 * OpenAPI spec version: 0.1.0
 */
import type { CourseLevel } from './courseLevel';

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  instructor?: string | null;
  provider: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  duration?: string | null;
  level: CourseLevel;
  tags?: string[];
  isFree?: boolean;
  createdAt: Date;
}